requirements.txt
langchain==0.1.1
pinecone-client==3.1.0
streamlit==1.33.0
tiktoken==0.4.0
openai==1.20.0
pypdf==4.2.0